requirements.txt
argcomplete==1.12.3
asgiref==3.5.2
autopep8==1.7.0
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
colorama==0.4.5
cryptography==37.0.4
decli==0.5.2
distlib==0.3.6
Django==3.2.15
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
docutils==0.19
dotty-dict==1.3.1
filelock==3.8.0
gitdb==4.0.9
GitPython==3.1.27
identify==2.5.3
idna==3.3
importlib-metadata==4.12.0
invoke==1.7.1
jaraco.classes==3.2.2
jeepney==0.8.0
Jinja2==3.1.2
keyring==23.9.0
Markdown==3.4.1
MarkupSafe==2.1.1
more-itertools==8.14.0
nodeenv==1.7.0
packaging==21.3
pkginfo==1.8.3
platformdirs==2.5.2
pre-commit==2.20.0
prompt-toolkit==3.0.31
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
PyJWT==1.7.1
pyparsing==3.0.9
python-gitlab==3.9.0
python-semantic-release==7.31.4
pytz==2022.2.1
PyYAML==6.0
questionary==1.10.0
readme-renderer==37.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
SecretStorage==3.3.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
termcolor==1.1.0
toml==0.10.2
tomlkit==0.11.4
tqdm==4.64.1
twine==3.8.0
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.4
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.8.1
